Update Docker command options in the GitHub Actions workflow
Update the command syntax used to run the built Docker image in the
GitHub Actions workflow to correctly support cgroups v2. This is
necessary because as of Ubuntu 22.04 (which is being rolled out for
workflows using `ubuntu-latest` runners) cgroups v2 is used. The
default namespace in v2 is `private` whereas on v1 it was `host`. We
must specify the `host` namespace to get the same behavior as seen
previously.